Update [in_memory.rs (mod tests)]

+++ src/engine/in_memory.rs
@@ -174,7 +174,7 @@ mod tests {
         collection.insert_batch(batch).unwrap();
         let duration_insert = start_insert.elapsed();
         println!("✅ Finished inserting {} vectors in {:?}", num_vectors, duration_insert);
-        println!("   Average insert time: {:?}", duration_insert / num_vectors);
+        println!("   Average insert time: {:?}", duration_insert / num_vectors as u32);
 
         println!("\n🔍 Running queries...");
         let num_queries = 100;
